pyproject.toml
[project]
name = "wbep"
version = "0.1.0"
description = ""
authors = [
    {name = "Vitaliy Dedhorg",email = "qwertyone120@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
